增加用户注册功能

+++ app.py
@@ -1,24 +1,22 @@
 # -*- coding:utf-8 -*-
 import os
-from datetime import datetime
 from threading import Thread
 
-from flask import Flask, request, current_app
+from flask import Flask, request
 from flask import redirect, url_for
 from flask import flash
 from flask import render_template
 
-from flask_login import login_user, LoginManager, UserMixin, login_required, logout_user, current_user
+from flask_login import login_user, LoginManager, login_required, logout_user, current_user
 from flask_bootstrap import Bootstrap
 from flask_script import Manager, Shell
-from werkzeug.security import generate_password_hash, check_password_hash
-from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate, MigrateCommand
 from flask_moment import Moment
 from flask_mail import Mail, Message
-from itsdangerous import TimedJSONWebSignatureSerializer as Serializer
 
 from form import LoginForm, TaskForm, RegistrationForm
+from models import db, User, Task
+
 
 basedir = os.path.abspath(os.path.dirname(__file__))
 
@@ -29,18 +27,18 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + os.path.join(basedir, 'da
 app.config['SQLALCHEMY_COMMIT_ON_TEARDOWN'] = True
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 
-app.config['MAIL_SERVER'] = 'smtp.126.com'  # kidult1107@126.com，使用这个的
-app.config['MAIL_PORT'] = 465
-app.config['MAIL_USE_TLS'] = True
+app.config['MAIL_SERVER'] = 'smtp.126.com'
+app.config['MAIL_PORT'] = 25
+app.config['MAIL_USE_TLS'] = False
 app.config['MAIL_USERNAME'] = os.environ.get('MAIL_USERNAME')
 app.config['MAIL_PASSWORD'] = os.environ.get('MAIL_PASSWORD')
-app.config['TODOLIST_MAIL_SUBJECT_PREFIX'] = '[TODOLIST]'
-app.config['TODOLIST_MAIL_SENDER'] = 'TODOLIST Admin <kidult1107@126.com>'
-app.config['TODOLIST_ADMIN'] = os.environ.get('TODOLIST_ADMIN')  # Admin管理者邮箱
+app.config['FLASKY_MAIL_SUBJECT_PREFIX'] = '[TODOLIST]'
+app.config['FLASKY_MAIL_SENDER'] = 'Daytodo-Admin<kidult1107@126.com>'
+app.config['FLASKY_ADMIN'] = os.environ.get('FLASKY_ADMIN')  # Admin管理者邮箱
 
 manager = Manager(app)
 bootstrap = Bootstrap(app)
-db = SQLAlchemy(app)
+db.init_app(app)
 migrate = Migrate(app, db)
 moment = Moment(app)
 mail = Mail(app)
@@ -60,8 +58,8 @@ def send_async_email(app, msg):
 
 
 def send_email(to, subject, template, **kwargs):
-    msg = Message(app.config['TODOLIST_MAIL_SUBJECT_PREFIX'] + ' ' + subject,
-                  sender=app.config['TODOLIST_MAIL_SENDER'], recipients=[to])
+    msg = Message(app.config['FLASKY_MAIL_SUBJECT_PREFIX'] + ' ' + subject,
+                  sender=app.config['FLASKY_MAIL_SENDER'], recipients=[to])
     msg.body = render_template(template + '.txt', **kwargs)
     msg.html = render_template(template + '.html', **kwargs)
     thr = Thread(target=send_async_email, args=[app, msg])
@@ -69,78 +67,16 @@ def send_email(to, subject, template, **kwargs):
     return thr
 
 
-class User(UserMixin, db.Model):
-    """
-    UserMixin 实现了使用flask_login插件所需要调用的方法。
-    """
-    __tablename__ = "users"
-    id = db.Column(db.Integer, primary_key=True)
-    email = db.Column(db.String(64), unique=True, index=True)
-    username = db.Column(db.String(64), unique=True, index=True)
-    password_hash = db.Column(db.String(128))
-    confirmed = db.Column(db.Boolean, default=False)
-
-    @property
-    def password(self):
-        raise AttributeError('password is not a readable attribute')
-
-    @password.setter
-    def password(self, password):
-        self.password_hash = generate_password_hash(password)
-
-    def verify_password(self, password):
-        return check_password_hash(self.password_hash, password)
-
-    def generate_confirmation_token(self, expiration=3600):
-        s = Serializer(current_app.config['SECRET_KEY'], expiration)
-        return s.dumps({'confirm': self.id}).decode('utf-8')
-
-    def confirm(self, token):
-        s = Serializer(current_app.config['SECRET_KEY'])
-        try:
-            data = s.loads(token.encode('utf-8'))
-        except:
-            return False
-        if data.get('confirm') != self.id:
-            return False
-        self.confirmed = True
-        db.session.add(self)
-        db.session.commit()
-        return True
-
-    def __init__(self, email, username, password):
-        self.email = email
-        self.username = username
-        self.password = generate_password_hash(password)
-
-
-class Task(db.Model):
-    __tablename__ = 'tasks'
-    id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, nullable=False)  # user外键，关联User表，回头可通过userid，找到对应user的task
-    title = db.Column(db.String(2048), nullable=False)
-    status = db.Column(db.Integer, nullable=False)  # 标志该项任务是否完成， 0 代表未完成， 1代表完成
-    create_time = db.Column(db.DateTime, nullable=False)  # 创建时间，使用datetime.datetime类型，可用于utc计时
-
-    def __init__(self, user_id, title, status=0):
-        self.user_id = user_id
-        self.title = title
-        self.status = status
-        self.create_time = datetime.utcnow()
-
-
-@app.before_request
-def before_request():
-    if current_user.is_authenticated \
-            and not current_user.confirmed \
-            and request.endpoint \
-            and request.endpoint != 'static':
-        return redirect(url_for('unconfirmed'))
+# 无法使用这个，使用后会报错
+# @app.before_request
+# def before_request():
+#     if not current_user.confirmed:
+#         return redirect(url_for('unconfirmed'))
 
 
 @app.route('/unconfirmed')
 def unconfirmed():
-    if current_user.is_anonymous or current_user.confirmed:
+    if current_user.confirmed:
         return redirect(url_for('index'))
     return render_template('unconfirmed.html')
 
@@ -148,6 +84,8 @@ def unconfirmed():
 @app.route("/")
 @login_required
 def index():
+    if not current_user.confirmed:
+        return redirect(url_for('unconfirmed'))
     # 使用current_user.id查询到对应的task，并使用模版渲染
     tasks = Task.query.filter_by(user_id=current_user.id).all()
     return render_template("index.html", tasks=tasks)
@@ -156,6 +94,8 @@ def index():
 @app.route("/add", methods=["GET", "POST"])
 @login_required
 def add():
+    if not current_user.confirmed:
+        return redirect(url_for('unconfirmed'))
     form = TaskForm()
     if form.validate_on_submit():
         task = Task(current_user.id, form.title.data, form.status.data)
@@ -169,6 +109,8 @@ def add():
 @app.route("/delete/<int:article_id>")
 @login_required
 def delete(article_id):
+    if not current_user.confirmed:
+        return redirect(url_for('unconfirmed'))
     task = Task.query.filter_by(id=article_id).first_or_404()
     db.session.delete(task)
     db.session.commit()
@@ -179,6 +121,8 @@ def delete(article_id):
 @app.route("/change/<int:article_id>", methods=["GET", "POST"])
 @login_required
 def change(article_id):
+    if not current_user.confirmed:
+        return redirect(url_for('unconfirmed'))
     form = TaskForm()
     task = Task.query.filter_by(id=article_id).first_or_404()
     if form.validate_on_submit():
@@ -197,8 +141,6 @@ def login():
     form = LoginForm()
     if form.validate_on_submit():
         user = User.query.filter_by(email=form.email.data).first()
-        # print("form username, form password:", form.username.data, form.password.data)
-        # print("username, password", user.username, user.password_hash)
         if user is not None and user.verify_password(form.password.data):
             login_user(user)
             return redirect(url_for("index"))
@@ -233,9 +175,9 @@ def register():
 
 
 @app.route('/confirm/<token>')
-@login_required  # 有点疑问，此时从邮箱点击这个，能登录吗？
+#@login_required
 def confirm(token):
-    if current_user.confirmed:  # 有点疑问，此时从邮箱点击这个，能登录吗？有current_user这个信息吗？
+    if current_user.confirmed:  # 有点疑问，从邮箱点击这个，此时有current_user这个信息吗？
         return redirect(url_for('index'))
     if current_user.confirm(token):
         flash('You have confirmed your account. Thanks!')
